Add support for new signature in execute
The new signature puts the output params at the end.

Since the number of output params depend on the number of params clad
is unable to know the number of output params during type deduction. To
solve this an overload for the grad function is created that has the same
number of output params as determined at type deduction stage (equal to the
number of params of the original function). The extra args are dummy
args and they are present just to match the DerivedFnType. The original
grad function is then called inside the overload function using the
required params.

Eg.
// Requested function:
int f(int a, int* b) {
  . . .
}

// Derived function:
void f_grad1(int a, int* b, int* d_b) {
  . . .
}

// Derived function overload:
void f_grad1(int a, int* b, int* d_b, int* d_0) {
  return f_grad1(a, b, d_b);
}